add parallax viewport to scale image depending on resolution
When providing a parallax image, a user may want to be able to influence
the image part that should be displayed. It would then provide an ideal
viewport for the image. If the viewport size matches the browser size,
nothing changes. However, if the browser size is smaller, the image will
be reduced to match the viewport and the reduction factor is also
applied to the steps. The ratio is kept and the scale factor is chosen
to "cover" the image (otherwise, parallax won't work).

This doesn't cover all possible situations, but it should be good enough
to handle resolution change without ratio change (1920×1080 → 1280×720)
or with a small ratio change (1920×1080 → 1024×768).

Fix: #855